split the templated Sources into its own files
